+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## main (unreleased)
 
-- Add support for `in-ns` forms in `clojure-ts-find-ns`
+- [#38]: Add support for `in-ns` forms in `clojure-ts-find-ns`.
 
 ## 0.2.2 (2024-02-16)