+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [#46]: Fix missing `comment-add` variable in `clojure-ts-mode-variables` mentioned in [#26]
 - Add imenu support for `deftest` definitions.
 - [#53]: Let `clojure-ts-mode` derive from `clojure-mode` for Emacs 30+.
+- [#42]: Fix imenu support for definitions with metadata.
 
 ## 0.2.2 (2024-02-16)